[Fix/#21] 1차 코드 리뷰 반영
